Implement keyboard-based AI and updated camera movement
Enemy can now be controlled by keyboard for debugging purposes. Improved the arena scrolling to be more resilient to mechs being out of bounds. Can still be broken when mechs move while the arena is scrolling - should implement a newpos for mech to check arena bounds against a future position.