Fix state cyclic dependency
When a state was returning another state to transition to, it would create a cyclic dependency when compiling, which godot does not allow. Instead of returning a state, the state now returns a string with the new state to transition to, which the AI class uses to switch to the new state.